@@ -15,14 +15,17 @@
   },
   "license": "GPL-3.0",
   "engines": {
-    "node": ">=8.12.0 <9.0.0",
+    "node": ">=10.2.1 <11.0.0",
     "npm": ">=6.4.0 <7.0.0"
   },
   "devDependencies": {
+    "@types/decompress": "4.2.0",
     "@types/electron-devtools-installer": "2.0.2",
     "@types/electron-settings": "3.0.0",
+    "@types/file-saver": "0.0.1",
     "@types/fs-extra": "5.0.1",
-    "@types/jasmine": "2.8.8",
+    "@types/jasmine": "3.3.12",
+    "@types/jasminewd2": "2.0.3",
     "@types/jquery": "3.3.29",
     "@types/jsonfile": "4.0.1",
     "@types/lodash": "4.14.121",
@@ -30,7 +33,7 @@
     "@types/node-hid": "0.7.0",
     "@types/request": "2.0.8",
     "@types/semver": "5.5.0",
-    "@types/usb": "1.1.3",
+    "@types/tmp": "0.0.33",
     "autoprefixer": "6.5.3",
     "buffer": "5.0.6",
     "check-node-version": "^3.2.0",
@@ -41,12 +44,12 @@
     "decompress": "4.2.0",
     "decompress-tarbz2": "4.1.1",
     "devtron": "1.4.0",
-    "electron": "2.0.16",
+    "electron": "3.1.10",
     "electron-builder": "20.34.0",
     "electron-debug": "1.5.0",
     "electron-devtools-installer": "2.2.3",
     "electron-log": "2.2.16",
-    "electron-rebuild": "1.8.2",
+    "electron-rebuild": "1.8.5",
     "electron-settings": "3.1.4",
     "electron-updater": "2.21.4",
     "exports-loader": "0.6.3",
@@ -54,15 +57,15 @@
     "fs-extra": "5.0.0",
     "gh-pages": "2.0.1",
     "html-webpack-plugin": "3.2.0",
-    "jasmine": "2.8.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-node": "2.0.1",
+    "jasmine": "3.4.0",
+    "jasmine-core": "3.4.0",
+    "jasmine-node": "3.0.0",
     "jasmine-ts": "0.3.0",
     "jsonfile": "4.0.0",
-    "lerna": "3.13.0",
+    "lerna": "3.14.1",
     "lodash": "4.17.11",
     "mkdirp": "0.5.1",
-    "node-hid": "0.7.3",
+    "node-hid": "0.7.8",
     "npm-run-all": "4.0.2",
     "nrf-intel-hex": "1.3.0",
     "postcss-url": "8.0.0",
@@ -70,7 +73,7 @@
     "request": "2.88.0",
     "rimraf": "2.6.1",
     "source-map-support": "0.5.9",
-    "stylelint": "9.6.0",
+    "stylelint": "10.0.0",
     "svg-sprite": "1.5.0",
     "ts-loader": "5.3.3",
     "ts-node": "7.0.1",
@@ -100,6 +103,5 @@
     "clean": "lerna exec rimraf ./node_modules ./dist && rimraf ./node_modules ./dist ./tmp",
     "predeploy-gh-pages": "lerna run build:web --scope=uhk-web",
     "deploy-gh-pages": "gh-pages -d packages/uhk-web/dist"
-  },
-  "dependencies": {}
+  }
 }
