@@ -7,5 +7,5 @@ const device = uhk.getUhkDevice();
 setInterval(() => {
     areLedsEnabled = !areLedsEnabled;
 
-    device.write(uhk.getTransferData(new Buffer([uhk.usbCommands.setTestLed, areLedsEnabled ? 1 : 0])));
+    device.write(uhk.getTransferData(Buffer.from([uhk.usbCommands.setTestLed, areLedsEnabled ? 1 : 0])));
 }, 500);
