@@ -3,7 +3,7 @@ const uhk = require('./uhk');
 const device = uhk.getUhkDevice();
 
 function readKeyboardState() {
-    const payload = new Buffer([uhk.usbCommands.getDeviceState]);
+    const payload = Buffer.from([uhk.usbCommands.getDeviceState]);
     console.log('Sending ', uhk.bufferToString(payload));
     device.write(uhk.getTransferData(payload));
     const receivedBuffer = device.readSync();
